# openapi_contract.json
{
  "components": {
    "schemas": {
...
      "AuthLoginIn": {
        "properties": {
          "email": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "mfa_code": {
            "maxLength": 12,
            "minLength": 6,
            "title": "Mfa Code",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "tenant_slug": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Tenant Slug",
            "type": "string"
          }
        },
        "required": [
          "tenant_slug",
          "email",
          "password"
        ],
        "title": "AuthLoginIn",
        "type": "object"
      },
      "AuthLogoutOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "AuthLogoutOut",
        "type": "object"
      },
...
      "AuthRefreshIn": {
        "properties": {
          "refresh_token": {
            "maxLength": 300,
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          },
          "tenant_slug": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Tenant Slug",
            "type": "string"
          }
        },
        "required": [
          "tenant_slug",
          "refresh_token"
        ],
        "title": "AuthRefreshIn",
        "type": "object"
      },
...
      "AuthTokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "mfa_enabled": {
            "title": "Mfa Enabled",
            "type": "boolean"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "tenant_slug": {
            "title": "Tenant Slug",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in_seconds",
          "tenant_slug",
          "role",
          "mfa_enabled"
        ],
        "title": "AuthTokenOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Telegram-driven salon management API",
    "title": "SalonOS",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLogoutOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}